+++ ios/AppLovinMAXNativeAdView.m
@@ -305,21 +305,10 @@
     nativeAdInfo[@"callToAction"] = ad.callToAction;
     nativeAdInfo[@"starRating"] = ad.starRating;
     
-    if ( !isnan(ad.mediaContentAspectRatio) )
+    // The aspect ratio can be 0.0f when it is not provided by the network.
+    if ( ad.mediaContentAspectRatio > 0 )
     {
-        // The aspect ratio can be 0.0f when it is not provided by the network.
-        if ( fabs(ad.mediaContentAspectRatio) < FLT_EPSILON )
-        {
-            nativeAdInfo[@"mediaContentAspectRatio"] = @(1.0);
-        }
-        else
-        {
-            nativeAdInfo[@"mediaContentAspectRatio"] = @(ad.mediaContentAspectRatio);
-        }
-    }
-    else
-    {
-        nativeAdInfo[@"mediaContentAspectRatio"] = @(1.0);
+        nativeAdInfo[@"mediaContentAspectRatio"] = @(ad.mediaContentAspectRatio);
     }
     
     nativeAdInfo[@"isIconImageAvailable"] = @(ad.icon != nil);