+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Versions
 
 ## x.x.x
-* Remove the default aspect ratio of a media view from the native ad properties.
+* Remove the default media view aspect ratio of `1.0` when a mediated network does not provide it.
 * Fix IconView not being displayed when setting it up with an empty DOM node.
 ## 5.2.3
 * Fix banner/MREC crashes related to previous release.