+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## x.x.x
 * Fix InMobi media view not showing due to lack of sizing.
+## 5.3.0
+* Remove the default media view aspect ratio of `1.0` when a mediated network does not provide it.
+* Fix LINE media view not showing.
 * Fix IconView not being displayed when setting it up with an empty DOM node.
 ## 5.2.3
 * Fix banner/MREC crashes related to previous release.